+++ src/Sentry.Unity/ScriptableSentryUnityOptions.cs
@@ -2,7 +2,6 @@ using System;
 using Sentry.Extensibility;
 using Sentry.Unity.Integrations;
 using UnityEngine;
-using UnityEngine.Serialization;
 
 namespace Sentry.Unity
 {
@@ -28,7 +27,11 @@ namespace Sentry.Unity
 
         [field: SerializeField] public string? Dsn { get; set; }
         [field: SerializeField] public bool CaptureInEditor { get; set; } = true;
+
         [field: SerializeField] public bool EnableLogDebouncing { get; set; } = false;
+        [field: SerializeField] public int DebounceTimeLog { get; set; } = (int)TimeSpan.FromSeconds(1).TotalMilliseconds;
+        [field: SerializeField] public int DebounceTimeWarning { get; set; } = (int)TimeSpan.FromSeconds(1).TotalMilliseconds;
+        [field: SerializeField] public int DebounceTimeError { get; set; } = (int)TimeSpan.FromSeconds(1).TotalMilliseconds;
 
         [field: SerializeField] public double TracesSampleRate { get; set; } = 0;
         [field: SerializeField] public bool AutoStartupTraces { get; set; } = true;
@@ -123,6 +126,9 @@ namespace Sentry.Unity
                 Dsn = Dsn,
                 CaptureInEditor = CaptureInEditor,
                 EnableLogDebouncing = EnableLogDebouncing,
+                DebounceTimeLog = TimeSpan.FromMilliseconds(DebounceTimeLog),
+                DebounceTimeWarning = TimeSpan.FromMilliseconds(DebounceTimeWarning),
+                DebounceTimeError = TimeSpan.FromMilliseconds(DebounceTimeError),
                 TracesSampleRate = TracesSampleRate,
                 AutoStartupTraces = AutoStartupTraces,
                 AutoSceneLoadTraces = AutoSceneLoadTraces,