+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Fixed the `BreadcrumbsForErrors` checkbox on the config window ([#1306](https://github.com/getsentry/sentry-unity/pull/1306))
 - The SDK filters `Bad Gateway` Exceptions of type `Exception` by default ([#1293](https://github.com/getsentry/sentry-unity/pull/1293))
 
+### Features
+
+- Surfaced debounce times to the options ([#1310](https://github.com/getsentry/sentry-unity/pull/1310))
+
 ### Dependencies
 
 - Bump CLI from v2.16.1 to v2.17.5 ([#1288](https://github.com/getsentry/sentry-unity/pull/1288), [#1289](https://github.com/getsentry/sentry-unity/pull/1289), [#1299](https://github.com/getsentry/sentry-unity/pull/1299))